pyproject.toml
[tool.poetry]
name = "alembic-partition"
version = "0.1.0"
description = ""
authors = ["SmirnovaT"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alembic = "^1.13.2"
sqlalchemy = "^2.0.31"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.4"
psycopg2-binary = "^2.9.9"
pre-commit = "^3.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
